Adding the multi_subset_bam option
A tool developed by Stefan Lang to subset a 10X bam file per barcode (part of the Rustody package). The multi_subset_bam runs with Rust and takes several tsvs as input (as a difference to 10X subset-bam tool). Given that the heaviest part of the operation is to read and write bam files, this simple change significantly speeds up the process of subsetting a set of cell barcodes.